Add socket connection logic
