+++ middleware/socket.js
@@ -1,0 +1,42 @@
+const socket = require("socket.io");
+
+const configureSocket = (server) => {
+  const io = socket(server, {
+    cors: {
+      origin: "http://localhost:3002",
+      methods: ["GET", "POST"],
+    },
+  });
+
+  const smartWatchConnections = new Map();
+  const dashboardConnections = new Map();
+
+  io.on("connection", (socket) => {
+    socket.on("connectSmartWatch", async (patientId) => {
+      smartWatchConnections.set(patientId, socket);
+    });
+
+    socket.on("connectDashboard", (patientId) => {
+      dashboardConnections.set(patientId, socket);
+    });
+
+    socket.on("watchData", (vitals) => {
+      const patientId = vitals["patientId"];
+      const dashboardSocket = dashboardConnections.get(
+        "651a8404d62a21c7687b89a8"
+      );
+      if (dashboardSocket) {
+        dashboardSocket.emit("updateVitals", vitals);
+      } else {
+        console.log(`No dashboard found for patient ID ${patientId}`);
+      }
+    });
+
+    socket.on("disconnect", () => {
+      smartWatchConnections.delete(socket.id);
+      dashboardConnections.delete(socket.id);
+    });
+  });
+};
+
+module.exports = configureSocket;
